requirements.txt
Django==4.2.7
django-allauth==0.57.0
Pillow==10.0.1
PyMuPDF==1.23.8
python-docx==0.8.11
pytesseract==0.3.10
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
faiss-cpu==1.7.4
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
reportlab==4.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
